requirements.txt
asgiref==3.8.1
backports.tarfile==1.2.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
crashtest==0.4.1
cryptography==44.0.0
distlib==0.3.9
Django==5.1.4
django-phonenumber-field==8.0.0
dulwich==0.22.7
fastjsonschema==2.21.1
filelock==3.16.1
idna==3.10
importlib_metadata==8.5.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
keyring==25.6.0
more-itertools==10.5.0
msgpack==1.1.0
packaging==24.2
phonenumbers==8.13.52
pkginfo==1.12.0
platformdirs==4.3.6
poetry==2.0.1
poetry-core==2.0.1
pycparser==2.22
pyproject_hooks==1.2.0
RapidFuzz==3.11.0
requests==2.32.3
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
sqlparse==0.5.3
tomlkit==0.13.2
trove-classifiers==2025.1.10.15
urllib3==2.3.0
virtualenv==20.28.1
zipp==3.21.0
